+++ src/bro/cli.py
@@ -5,8 +5,10 @@ import logging
 from pathlib import Path
 import sys
 
-from bro import reasoner, ui_io, executive
+from bro import ui_io
 from bro.reasoner import Context
+from bro.executive.openai_cua import OpenAiCuaExecutive
+from bro.reasoner.openai_generic import OpenAiGenericReasoner
 
 
 _logger = logging.getLogger(__name__)
@@ -19,13 +21,14 @@ def main() -> None:
     context = _build_context(sys.argv[1:])
 
     openai_api_key = os.getenv("OPENAI_API_KEY")
+
     ui = ui_io.make_controller()
-    exe = executive.OpenAiCuaExecutive(
+    exe = OpenAiCuaExecutive(
         ui=ui,
         state_dir=_dir,
         openai_api_key=openai_api_key,
     )
-    rsn = reasoner.OpenAiReasoner(
+    rsn = OpenAiGenericReasoner(
         executive=exe,
         ui=ui,
         state_dir=_dir,
