+++ src/bro/executive/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+from abc import ABC, abstractmethod
+
+
+class Executive(ABC):
+    @abstractmethod
+    def act(self, goal: str) -> str:
+        raise NotImplementedError
